Avoid multiple consecutive spaces when copying verse
Previously, copying a verse could result in multiple
consecutive space characters if a part of the verse being
copied had a trailing space character and an additional
explicit space character was added in the next step
(due to 'spaceJustWritten' being 'false').

In order to avoid this, set the value of 'spaceJustWritten'
depending on the last character of the String just written.