Make the search activity keep state
Instead of calling finish() to return to the previous screen,
we start a new intent for the MainBibleActivity with FLAG_ACTIVITY_REORDER_TO_FRONT
to move the MainBibleActivity to the front and the SearchResults to the back.

Then upon pressing the back button, the SearchResults activity is moved to the front again.