Fix window button animiation issue when exiting search
The WindowInsets listener was firing because the MainActivity was in a paused
state, but had not yet stopped, when switching to the Search activity. This,
in combination with the keyboard opening immediately when startin gthe Search
trigged a resize event that was erroneously being caught by the MainActivity,
even though it wasn't visible. I added a variable to skip the listener
if the activty is paused.