Fix v11n conversion issue when converting verseranges with verse 0 included.
Solution is to skip verse 0 from conversions. If this is not desired in some case,
we should make exception there, but normally verse 0 is not desired to be in verse ranges because
it always is converted to corresponding 0-verse in another versification, which leads to
weird verseranges (can be even with end larger than start).