In sync with workspace icon methodology changed
The method I have implemented shows the Workspace icon in green when the window is using the workspace setting and no overlay icon at all when it is a window setting. This makes more sense to me, simplifies the menu a little, and visually ties everything it back to the workspace concept.