Changed the style of bar buttons for extra clarity
The unpinned buttons looked the same as the pinned buttons (with the exception of tiny pin icon). The unpinned buttons function as a group so it seemed better to display them so that the look like a group and distinct from the other buttons. So i reduced the corners a little. I tried hard to not have to create new resource files for them but alas, I could not see how to manipulate just the corner style.