Added maximise button to the verse dialog
I think this will be really handy. It allows the user to maximise the current window with a single click. Great when entering notes especially. But also useful when adding bookmarks if there are more than a few windows on screen.

Unfortunately it doesn't work properly. If you click the button twice the other windows go white. And if you then fullscreen it the maximised window is displayed correctly. I suspect it is not too hard to fix for someone who knows what he is doing. I am guessing that it must be the context from which i am calling the action - i am just calling the existing maximise function.