Use gradle's configuration-avoidance APIs to add dependencies
Instead of causing the tasks to be configured immediately in order to
inject a dependency on the custom tasks, use gradle's recommended API to
delay the task's configuration.
This also prepares the groovy code for migration to kotlin, because the
"preBuild" and "check" tasks are not declared as variables in scope.