Workaround duplicate classes from transitive dependencies
Reference: https://github.com/android/android-test/issues/861

The transitive dependencies of espresso-accessibility and
espresso-contrib include versions of both "checker", and "checker-qual"
which contains a portion of "checker".  This results in duplicate
classes which cause the build of "androidTest" variants to fail.

This commit applies the workaround, described in the referenced github
issue, of excluding "checker" from the transitive dependencies.