Launch in the activity's lifecycleScope, not GlobalScope
This will ensure that the job is canceled when the activity is
destroyed.  Jobs that are launched in the GlobalScope are leaked if they
do not complete and are not otherwise explicitly canceled.
This job is leaked in the GlobalScope when the device is rotated while
the confirmation dialog is displayed, because it neither completes nor
is canceled.  When launched in the activity's lifecycleScope, it is
canceled when the activity is destroyed, due to device rotation, while
the confirmation dialog is displayed.